requirements.txt
numpy
Pillow
matplotlib
opencv-python
scikit-image
scipy
scikit-learn
seaborn